Eight Winds Engine version 1.1.0
Major Changes:

    Boost dependency was removed
    EWETexture has been rewritten and broken into a few other classes.
    Quite a few changes to the file structure.
        Models and texture classes were moved into a sub folder.
        The Skin, Static, and Rigid rendering systems have been moved. Static was renamed to stationary, and is still incomplete.
        UIComponents were split into a few classes
    imgui was brought back into the engine.
    The cmake file was reworked. Simplified the process of getting the dependencies

+++ src/LevelBuilderMM.cpp
@@ -1,4 +1,4 @@
-#include "EWEngine/gui/LevelBuilderMM.h"
+#include "EWEngine/GUI/LevelBuilderMM.h"
 
 namespace EWE {
 	LevelBuilderMM::LevelBuilderMM(float screenWidth, float screenHeight, GLFWwindow* windowPtr) {
@@ -58,7 +58,7 @@ namespace EWE {
 	}
 
 	void LevelBuilderMM::processClick(double xpos, double ypos) {
-		std::pair<UIComponentTypes, int8_t> returnValues = MenuModule::checkClick(xpos, ypos);
+		std::pair<UIComponentTypes, int16_t> returnValues = MenuModule::checkClick(xpos, ypos);
 		
 
 		//return MCR_none;
